+++ esep_ss2021_Gruppe2_3/src/logic/datamodel/encoded_workpiece.h
@@ -27,7 +27,7 @@ public:
     EncodedWorkpiece(uint32_t code);
 
     id get_id();
-    float get_height();
+    int get_height();
     WorkpieceType get_type();
 };
